Move RuntimeMock to its own file
