+++ tests/php/RuntimeTest.php
@@ -1,43 +1,14 @@
 <?php
 
-namespace Phel;
+declare(strict_types=1);
+
+namespace PhelTest;
 
 use PHPUnit\Framework\TestCase;
 
-class RuntimeMock extends Runtime
+final class RuntimeTest extends TestCase
 {
-    public $files = [];
-    public $loadedFile = null;
-
-    public function __construct()
-    {
-        $this->globalEnv = new GlobalEnvironment();
-    }
-
-    public function setFiles(array $files)
-    {
-        $this->files = $files;
-    }
-
-    protected function loadFile(string $filename, string $ns): bool
-    {
-        $this->loadedFile = $filename;
-        return true;
-    }
-
-    protected function fileExists($filename): bool
-    {
-        return in_array($filename, $this->files);
-    }
-}
-
-class RuntimeTest extends TestCase
-{
-
-    /**
-     * @var RuntimeMock
-     */
-    protected $runtime;
+    private RuntimeMock $runtime;
 
     public function setUp(): void
     {
@@ -57,7 +28,7 @@ class RuntimeTest extends TestCase
         $this->runtime->addPath('Foo\\Bar\\Baz\\Dib\\Zim\\Gir\\', ['/vendor/foo.bar.baz.dib.zim.gir/src']);
     }
 
-    public function testExistingFile()
+    public function testExistingFile(): void
     {
         $this->runtime->loadNs('Foo\Bar\Core');
         $this->assertEquals(
@@ -72,12 +43,12 @@ class RuntimeTest extends TestCase
         );
     }
 
-    public function testMissing()
+    public function testMissing(): void
     {
         $this->assertFalse($this->runtime->loadNs('No_Vendor\No_Package\NoClass'));
     }
 
-    public function testDeepFile()
+    public function testDeepFile(): void
     {
         $this->runtime->loadNs('Foo\Bar\Baz\Dib\Zim\Gir\Core');
         $this->assertEquals(
@@ -86,7 +57,7 @@ class RuntimeTest extends TestCase
         );
     }
 
-    public function testConfusion()
+    public function testConfusion(): void
     {
         $this->runtime->loadNs('Foo\Bar\Doom');
         $this->assertEquals(
